Fixing Consumption ratio

See merge request bppetrokimiagresik/akutansi-biaya/akutansibiaya!86

+++ app/Http/Controllers/PriceRenDaanController.php
@@ -13,6 +13,7 @@ use App\Models\QtyRenDaan;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 use Maatwebsite\Excel\Facades\Excel;
 
 class PriceRenDaanController extends Controller
@@ -28,13 +29,16 @@ class PriceRenDaanController extends Controller
     public function create(Request $request)
     {
         try {
-            $request->validate([
+            $validator = Validator::make($request->all(), [
                 "version_asumsi" => 'required',
                 "bulan" => 'required',
                 "material_id" => 'required',
                 "region_id" => 'required',
                 "price_rendaan_value" => 'required',
-            ]);
+            ], validatorMsg());
+
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
 
 
             $input['version_id'] = $request->version_asumsi;
@@ -56,13 +60,17 @@ class PriceRenDaanController extends Controller
     public function update(Request $request)
     {
         try {
-            $request->validate([
+
+            $validator = Validator::make($request->all(), [
                 "version_asumsi" => 'required',
                 "bulan" => 'required',
                 "material_id" => 'required',
                 "region_id" => 'required',
                 "price_rendaan_value" => 'required',
-            ]);
+            ], validatorMsg());
+
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
 
             $input['version_id'] = $request->version_asumsi;
             $input['asumsi_umum_id'] = $request->bulan;
@@ -104,12 +112,15 @@ class PriceRenDaanController extends Controller
     public function import(Request $request)
     {
         try {
-            DB::transaction(function () use ($request){
-                $request->validate([
-                    'file' => 'required',
-                    'version' => 'required'
-                ]);
+            $validator = Validator::make($request->all(), [
+                "file" => 'required',
+                "version" => 'required',
+            ], validatorMsg());
 
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
+
+            DB::transaction(function () use ($request){
                 PriceRenDaan::where('version_id', $request->version_id)->delete();
 
                 $file = $request->file('file')->store('import');
